requirements.txt
absl-py==0.9.0
asgiref==3.2.7
boto3==1.14.3
botocore==1.17.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-storages==1.9.1
django-tinymce4-lite==1.8.0
djangorestframework==3.11.0
docutils==0.15.2
fanstatic==1.1
gast==0.3.3
google-api-core==1.20.0
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-pasta==0.2.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.29.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
jmespath==0.10.0
js==1.0
jsmin==2.2.2
Markdown==3.2.2
node==0.9.25
numpy==1.18.4
oauthlib==3.1.0
odict==1.7.0
Pillow==7.1.2
plumber==1.5
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
shutilwhich==1.1.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
WebOb==1.8.6
Werkzeug==1.0.1
whitenoise==5.1.0
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.1.0
zope.lifecycleevent==4.3
zope.proxy==4.3.5
